+++ packages/conc/lib/Polysemy/Conc.hs
@@ -54,6 +54,7 @@ module Polysemy.Conc (
   timeoutAs_,
   timeoutU,
   timeoutMaybe,
+  timeoutStop,
   retrying,
   retryingWithError,
 
@@ -267,7 +268,7 @@ import Polysemy.Conc.Interpreter.SyncRead (syncRead)
 import Polysemy.Conc.Monitor (ClockSkewConfig (ClockSkewConfig), clockSkewConfig, monitorClockSkew)
 import Polysemy.Conc.Queue (loop, loopOr)
 import Polysemy.Conc.Queue.Result (resultToMaybe)
-import Polysemy.Conc.Race (race_, timeoutAs, timeoutAs_, timeoutMaybe, timeoutU, timeout_)
+import Polysemy.Conc.Race (race_, timeoutAs, timeoutAs_, timeoutMaybe, timeoutStop, timeoutU, timeout_)
 import Polysemy.Conc.Retry (retrying, retryingWithError)
 import Polysemy.Conc.Sync (lock, withSync)