+++ packages/conc/lib/Polysemy/Conc/Race.hs
@@ -5,8 +5,9 @@ import Polysemy.Time (TimeUnit)
 
 import qualified Polysemy.Conc.Effect.Race as Race
 import Polysemy.Conc.Effect.Race (Race)
+import Polysemy.Resume (Stop, stop)
 
--- |Specialization of 'Race.race' for the case where both thunks return the same type, obviating the need for 'Either'.
+-- |Specialization of 'Race.race' for the case where both actions return the same type, obviating the need for 'Either'.
 race_ ::
   Member Race r =>
   Sem r a ->
@@ -16,8 +17,8 @@ race_ ml mr =
   unify <$> Race.race ml mr
 {-# inline race_ #-}
 
--- |Specialization of 'Race.timeout' for the case where the thunk return the same type as the fallback, obviating the
--- need for 'Either'.
+-- |Specialization of 'Race.timeout' for the case where the main action returns the same type as the fallback, obviating
+-- the need for 'Either'.
 timeout_ ::
   TimeUnit u =>
   Member Race r =>
@@ -41,7 +42,7 @@ timeoutAs err =
   Race.timeout (pure err)
 {-# inline timeoutAs #-}
 
--- |Specialization of 'timeoutAs' for the case where the thunk return the same type as the fallback, obviating the
+-- |Specialization of 'timeoutAs' for the case where the main action return the same type as the fallback, obviating the
 -- need for 'Either'.
 timeoutAs_ ::
   TimeUnit u =>
@@ -76,3 +77,14 @@ timeoutMaybe u ma =
   timeoutAs_ Nothing u (Just <$> ma)
 {-# inline timeoutMaybe #-}
 
+-- |Variant of 'Race.timeout' that calls 'Stop' with the supplied error when the action times out.
+timeoutStop ::
+  TimeUnit u =>
+  Members [Race, Stop err] r =>
+  err ->
+  u ->
+  Sem r a ->
+  Sem r a
+timeoutStop err =
+  timeout_ (stop err)
+{-# inline timeoutStop #-}